Refactor #21 - Add external persistence storage
- Clean code
- Fix package names
- Update project version
- Move Wallet extensions out of the interface
