+++ src/main/kotlin/com/rootsid/wal/library/dlt/model/Claim.kt
@@ -2,7 +2,6 @@ package com.rootsid.wal.library.dlt.model
 
 import io.iohk.atala.prism.api.CredentialClaim
 import io.iohk.atala.prism.identity.PrismDid
-import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.json.Json
@@ -25,7 +24,6 @@ data class Claim(
  *
  * Convert a Claim to PRISM CredentialClaim
  */
-@OptIn(ExperimentalSerializationApi::class)
 fun Claim.toCredentialClaim() = CredentialClaim(
     PrismDid.fromString(this.subjectDid),
     Json.decodeFromString(this.content)
