+++ src/test/kotlin/com/rootsid/wal/library/DLTKtTest.kt
@@ -1,7 +1,5 @@
 package com.rootsid.wal.library
 
-import org.junit.jupiter.api.Assertions.*
-
 internal class DLTKtTest {
 
     @org.junit.jupiter.api.Test
